pyproject.toml
[tool.poetry]
name = "assassin"
version = "0.1.0"
description = ""
authors = ["Timothy Morland <timothy.john.morland@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
Flask = "^2.2.2"
msal = "^1.20.0"
python-dotenv = "^0.21.0"
aiosql = "^7.1"
gunicorn = "^20.1.0"

[tool.poetry.group.dev.dependencies]
black = "^22.10.0"
isort = "^5.10.1"
pytest = "^7.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
